+++ CHANGELOG.md
@@ -6,12 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Handling missing ad unit's size
+### Removed
+- Ad units' tags from site details
 
 ## [1.18.5] - 2022-07-08
 ### Added
 - Checkbox for control size matching
-### Removed
-- Ad units' tags from site details
 
 ## [1.18.4] - 2022-06-30
 ### Added